// web/src/api/models/CacheRefreshConcurrency.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Admission behavior when a namespace has an unpublished refresh.
 */
export enum CacheRefreshConcurrency {
  REUSE = "reuse",
  CONFLICT = "conflict",
  PARALLEL = "parallel",
}